Drop deprecated properties in favor of Prettier
Stylelint has deprecated a slew of stylistic options in favor of switching to Prettier. In practice, Prettier covers most of the deprecated rules, with a few exceptions:

* `number-leading-zero`: forces leading zero
* `string-quotes`: does not enforce for `charset` strings
* `max-empty-lines`: sets to 1, which should be fine for most instances

Of those, the changes regarding `number-leading-zero` is the most problematic and relevant, as it goes against the Twitter Bootstrap style guide and Google.

The auto-wrapping behavior of Prettier is also annoying, as it prevents manual multi-line specification by the developer for visual clarity. This is less relevant in SCSS/CSS, however.